requirement.txt
Django == 4.1.4
djangorestframework == 3.13.1
mysqlclient
Pillow == 9.3.0
django-filter == 22.1
django-cors-headers == 3.13.0
livereload == 2.6.3
